Improved vector labeling
